fix(migrations): fix apply_migration INT4/INT8 type mismatch and dedup before unique index
- apply_migration decoded SELECT 1 as i64 (INT8) but Postgres returns INT4;
  change to i32 to fix startup crash on desk-server
- migration 009 failed with 'could not create unique index' on existing data
  with duplicate client_order_id values; prepend CTE-based dedup (keep highest
  id per user_id+client_order_id, null out the rest) before CREATE UNIQUE INDEX

+++ src/desk/db.rs
@@ -76,7 +76,7 @@ pub async fn run_migrations(pool: &PgPool) -> Result<(), sqlx::Error> {
 }
 
 async fn apply_migration(pool: &PgPool, version: &str, sql: &str) -> Result<(), sqlx::Error> {
-    let already_applied: Option<i64> =
+    let already_applied: Option<i32> =
         sqlx::query_scalar("SELECT 1 FROM _schema_migrations WHERE version=$1")
             .bind(version)
             .fetch_optional(pool)
